fix(mobile): move tab-screen tests out of app/ so expo-router doesn't load them as routes
Les fichiers app/(tabs)/{create,index}.test.tsx étaient traités comme des routes par
expo-router : leur jest.mock de top-level s'exécutait au runtime (ReferenceError:
Property 'jest' doesn't exist, red-box au démarrage) et ils polluaient le routage
(missing default export). Déplacés vers mobile/src/screens/ (convention du repo :
tous les autres tests vivent sous src/). Vérifié sur device (Galaxy A55) : plus de
red-box, écrans Créer/Voyages OK.
